+++ app/src/main/java/org/mozilla/fenix/library/bookmarks/BookmarkView.kt
@@ -16,7 +16,7 @@ import mozilla.components.support.base.feature.UserInteractionHandler
 import org.mozilla.fenix.NavGraphDirections
 import org.mozilla.fenix.R
 import org.mozilla.fenix.library.LibraryPageView
-import org.mozilla.fenix.library.SelectionInteractor
+import org.mozilla.fenix.selection.SelectionInteractor
 
 /**
  * Interface for the Bookmarks view.
